refactor: remove the dependency between consensus and ledger
