+++ crates/amaru-consensus/Cargo.toml
@@ -36,7 +36,6 @@ tracing.workspace = true
 tracing-opentelemetry.workspace = true
 
 amaru-kernel.workspace = true
-amaru-ledger.workspace = true
 amaru-mempool.workspace = true
 amaru-metrics.workspace = true
 amaru-observability.workspace = true
